display_device: support the "Not Charging" state
On Lenovo ThinkPads with the linux-next kernel, due to the
new kernel interface for the charging thresholds one battery could
appear as "pending-charge" in UPower thus breaking the DisplayDevice
that GNOME/MATE uses in the tray icon, displaying an unknown battery.

This patch adds a set of rules that the display device will follow:

1) If there is a SINGLE battery that is pending a charge, we will display
it as if the battery is full. That is because the state of
"pending-charge" is too complicated for the UI and could break it on
other desktops (such as Xfce).

2) If there is more than one battery but one is pending a charge, we
will return "Charging" because the other battery is charging.

None of the above will occur if there is no batteries or the state is
not unknown.

Signed-off-by: Ognjen Galic <smclt30p@gmail.com>